Final Submission
Final Submission

What does this code do?

You present the neural network with pre-defined dataset (this can be changed [max. 3 + bias]) calculates weights either given or generated at random, calculates the loss and hidden errors in hidden layers through number of epochs (iterations) with an outcome of probability distribution that hopefully justifies the confidence in the algorithm and output.